Added 1D-AE-KLI which works for small atoms with fully occupied subshells. I think it also works (up to the spherically averaged potential) for atoms with partially filled shells, but you cannot compare the results to real atoms since, the all_electron.py does it's calculation is with fractional occupation numbers and spin-symmetrically.
Is this enough for calculating setup's? The asymptiotical behaviour of 
KLI is -f_{HOMO}/r, (in B for example -0.16666/r) so it will create an 
error to at least the highest occupied orbital eigenvalues, '
but maybe this isn't so important since the these orbitals are not used anyway.

At some point, for example with Zn, the calcalation fails.
Maybe this is due to the missing gaunt's coefficients that I am
assuming to be zero.