-Renamed fermi_level to reference_level -Few comments added -Added possibility to add correlation to GLLB (P86) -Added possibility to use LUMO as a reference level instead of HOMO -Added possibility to "relax the core response" to new fermilevel (instead of using the one from setup-generator). Core orbitals and eigenvalues are still kept fixed in spirit of frozen core approximation.
I am working on a larger test suite to test all this, it is probably the best
that I won't add it to the standard test suite.

Intresting thing I discovered: 
For accurate atom HOMO eigenvalues, GLLB requires much larger cell (12Å)
compared to LDA, since LDA orbitals are more localized due to
exponential decay of exchange potential. The 3D eigenvalues in test/gllb2.py
will start to match better to 1D code if you do it with grid-spacing 0.18 and 12Å cell.