Initial work on the Allow-tests-to-be-run-in-parallel project.  The goal is that the tests sould run on at least 2 and 4 cores.
Another long-term goal is to move the test directory from trunk/test to trunk/gpaw/test (like for ASE) so that the tests are always available.  Like for ASE, one should do "cd /tmp; gpaw-test.py" to run the tests, and the tests should not be required to clean up files.