Discovered that my local tests were not running in the right installation.  This SHOULD fix the problems.
Please be good to me BuildBot.  Local tests pass,  hopefully nothing breaks.

Second attempt at making this actually work:

Added a "forces" convergence parameter.
It is only used if set to a value other than None,  otherwise it adds no overhead.
This feature should be considered somewhat experimental,  while it works functionally
its benefits are not well defined.  It appears to make relaxing the precision possible for
geometric relaxation as eventually the forces will normally settle to a reasonable precision
before the energies become accurate.
This feature has been documented and a test exists for it.