solvation: reduces copy and paste of code for zero width extrapolation
The zero width extrapolation is also used for the non-shortrange
("electrostatic") part of the total energy of a continuum solvent
calculation. In order to avoid divergence of the copy and pasted code
with the original GPAW calculator, a new method is added there which
takes care of the extrapolation. This method is now used by both GPAW
and SolvationGPAW.

Solvation tests are updated to check whether the electrostatic part is
calculated properly.