solvation: adds notes to copy and pasted c code
The multi grid solver has been extended to work with
nonlocal weights some time ago by copying and modifying parts of the c code.
This commit adds notes above the original copyright notes indicating
the origin of the copied files.

The weighted and unweighted versions should be merged together at some time
to avoid divergence of the two versions when changes are made to the
original multi grid code.