ENH: Gaussian is adapted to distance from border
The width of the Gaussian was estimated by the 
distance between center of box and the borders
and not the actual position of the Gaussian itself.

Also there is a minimum distance from the border 
a Gaussian should have (7 - 8 gps, see
https://listserv.fysik.dtu.dk/pipermail/gpaw-developers/2015-July/005806.html
)

Fixed, put charges to (inner) boundaries if
center of charge is between inner boundaries
and border. If center of charge is off box,
the coc will be centered. Warnings will be
raised in this two cases.