Improves accuracy of forces for solvation calculations.
The gradient of the dielectric is now calculated analytically
for the case of the effective potential method.
The finite differences which have been used before introduced
some noise on the cavity surface.
This is now fixed, such that forces are more accurate,
especially for very small molecules with hydrogen,
where the cavity can be steep.