Open pickle files in binary mode for python3
In python3, pickle expects to write and read to and from files in binary
mode, accessed by the `b' option to the ``open'' function. This commit
ensures that all files used for pickling have this option set.

The ``file'' function is also removed in favor of the ``open'' function,
even in the excercises