Make AEOrbital class work with lcao mode.
The expand method of the AEOrbital class assumes that certain wavefunction
arrays are populated. These are populated in grid mode and in plane wave mode,
but not in LCAO mode. The user can initialize the arrays in their own code, but
they have to be repopulated at every step in the self-consistent cycle. That's
what this patch does.

Foobar

Foobar