MAINT: react to delayed initialization of poisson solver
Due to recent changes, the solvation poisson solver is not
initialized (allocated) any more when it should print
information during GPAW startup.

In order to fix this, the allocation of the weighted FD
operators has been split into a separate method, allowing
the solvation poisson solver to instantiate the operators
early in the set_grid_descriptor method and allocate them later.

As a positive side effect, this splitting of initialization and
allocation of the weighted operators also removes a lot of code
duplication.