BUG: fix nan atomic radii passing silently in solvation model
If some of the atomic radii were set to `np.nan`, the output was
still finite. This commit fixes this by checking the radii
explicitly.

The fix is only a preliminary work-around to avoid unnoticed
wrong calculations. The root-cause of the problem is the too
generous masking of floating point errors with the analytical
limits in some expressions. This needs more attention and
will be fixed later.