Fix import errors found by flake8/pyflakes
Running flake8 on the GPAW code base revealed a number of pyflakes
errors. Two of the main issues were:

- Files which import functions/classes/variables and then do not use
  them.
- Files which used ``from x import *'', thus hiding any potential
  /missing/ imports.

For each of the ``imported but unused'' errors, the file has been
reviewed to determine whether the inclusion is truly an error. If it
is, the offending line has been deleted. However, not all of these
lines were errors. Some files imported objects to make them available
for export; in these cases, the objects have been added to the
``__all__'' variable. Other files imported objects to have them
available for eval. In these cases, the warning has been switched off
by the addition of a # noqa inline comment after the import.

All of the ``from x import *'' occurrences, have been removed and
explicitly replaced with the relevant imports.