Make QEH model use npz files, not pickle
Pickle files for storing numpy arrays are to be avoided - interoperability
between different versions of python is not guaranteed, and in principle opening
the pickle file could execute arbitrary code.

The pickle files in the QEH code were only being used to story numpy arrays, and
numpy has a file format exactly for this purpose, which can store a labelled
list of arrays in a single file. The files can then be opened using just the
numpy.load function, which gives a dictionary-type object and makes it possible
to refer to each array by the label.