BUG: Remove race conditions
When a building block is initialized, the last stage is to load the data (or as
much of it as exists) from a given file. When the remaining quantities are to be
calculated, one of the first steps is to save the current state to the same
file.

This can give rise to two different race conditions. Firstly, if a building
block is initialized, and the calculation is then performed, the master process
can potentially start writing to the file before all processes have finished
reading from it. Secondly, if a write were ever followed by a read, the other
nodes could start reading before the master had finished writing. The solution
to both problems is a world.barrier()

Fixes #90 - QEH tutorial fails