MAINT: moved attributes needed for Yukawa RSF to sep. object
To perform DSCF calculations utilizing RSF, Mg_pp has to be
caclulated "on the fly" (setup.calculate_yukawa_interaction).
Therefor "calculate_yukawa_interaction" and underlying methods
had to be integrated in BaseSetup. The corresponding attributes
had to be distributed over the inheriting classes. To not pollute
these classes too much, a new object/class is build, which
holds the according attributes.