Add option to calculate and store forces
When calculating the electron-phonon coupling, we are frequently interested in
calculating the phonons at the same time. The two are linked, in that they are
both computed based on self-consistent calculations on displaced structures. To
save computer time, it therefore makes sense to calculate both the forces and
the change in the potential after every self-consistent calculation.

Depends on ase !975