Add Easybuild configuration files.
These files are made for the foss/2019b and intel/2019b toolchains,
but are expected to work for future toolchains, too.  The
config-foss.py files will probably also work for other toolchains than
foss and intel, as long as it does not use Intel Math Kernel Library.

The special config file is necessary for the Intel toolchain as the
EasyBuild system reports FFT libraries that are build without -fPIC,
but we need to link to versions compiled *with* -fPIC when building
shared objects.