Add explicit flag for fixing the potential
The fix_potential flag prevents the evaluation of the potential on the
k-point grid of the band structure calculation.

At the moment, the flag is redundant as in the first iteration the
potential read from the file is used, and with fixed density
`update_potentials()` is not called anymore. However, the explicit flag
is supposed to clarify this behavior.