Change how parameters are structured.
Now everything for sj is in a single dictionary, and it uses the
GPAW and ASE set functions as much as possible rather than re-defining
everything. Also lots taken out of the __init__ since it was
redundant with set.

Still needs cleanup, and resulted in Issue #7 (re-initialization
of electronic structure stuff).