MAINT: Refactor plot_geom.py
The atoms are now plotted with the ASE plot_atoms function without loss of generality. This is useful when simulating larger molecules where it helps to have a different color/ size for each atom in the molecule. There are also a few cosmetic changes to make the nanosphere look golden.