Allow arbitrary installation path in gpaw-venv.sh
The gpaw-venv.sh for an installation of GPAW in a virtual
environment on Niflheim assumed that the install directory
was within the execution folder. This is really just a minimal
addition so an aritrary path can be used for installation.

Additionally a more intuitive way to change the repos has been
added.