Automatically write GC and C energy depending on calculation mode
I added an if statement that checks whether a target_potential
exists or not. If it does the GC energy is written otherwise the
canonical one. The keyword 'write_grandpotential_energy' overwrites
this automatic assignment. I tested various combinations of subsequent
'calc.set(...)' calls and it always seems to work. The only
exception is when the 'write_grandpotential_energy' is set before,
it is needed to be `set` to 'None' again to turn on the automatic
assignemnt again. I guess that is what we want anyway.