ENH: Print the applied constraints in the positions table of the log
Although used routinely and decisive for the final result, I noticed that
the applied (ASE-)constraints are not printed anywhere in the text output.
Hence, I created this suggestion where the constraints on the specific
atoms are printed right next to the atomic positions. The different
constraints get different labels and the labels are explained below the
table in a legend. The legend only contains the used constraints and does
not show if no constraints are used. Alternatively, the labels of the
constraints could also be given on the webpage instead of adding the legend.
Let me know what you think, I am more than happy to incorporate your
suggestions.