Transpose dipole moments in raman middle factor
I worked out by hand how to treat these.

As can be seen from the indices in the np.einsums, tN_ij, tN_xx,
and tN_mn are all supposed to be indexed by [n_from, n_to] for a
transition from electronic state n_from to state n_to.

Assuming that term1 was implemented correctly, then judging from
how the code sets t1_ij and t1_mn lines, it would appear to me that
the two operators are indexed as:

* elph_lnn[l, n_from, n_to]
* mom_dnn[d, n_to, n_from]

All of the rest of the tN_ij, tN_xx and tN_mn are implemented
consistently with this with the exception of four lines that
already had FIXMEs denoting the original author's uncertainty.
These lines require a simple transpose (no conjugate) so that they
are indexed by [n_from, n_to].

With this change, the output of GPAW on some gamma-only test data
for MoS2 finally matches my own implementation.